feat: move detailed movie card component to molecules folder
